+++ modules/providers/aws.js
@@ -6,7 +6,176 @@ var utils = require('../utils');
 
 var DEBUG = false;
 
+// http://cloud-images.ubuntu.com/releases/14.04/release/
+var DEFAULT_AMI = 'ami-64e27e0c'; // Ubuntu 14.04 64 bit, EBS
+var DEFAULT_REGION = 'us-east-1';
+var DEFAULT_SIZE = 't1.micro';
+
+exports.id = 'aws';
+exports.name = 'AWS';
+
 exports.pollDelay = 3000;
+
+// Provider interface
+
+exports.boot = function (instance, callback) {
+  var params = {
+    InstanceId: instance.aws.id,
+    region: instance.aws.region,
+    state: 'running'
+  };
+
+  exports.startInstance(params)
+    .then(exports.waitForInstanceState)
+    .then(exports.getFilteredInstances)
+    .catch(exports.catch)
+    .then(function (args) {
+      utils.updateInstance(instance.name, {
+        ip: args.Instances[0].PublicIpAddress,
+        aws: {
+          id: args.InstanceId,
+          size: args.Instances[0].InstanceType,
+          image: args.Instances[0].ImageId,
+          region: instance.aws.region,
+          monitoring: args.Instances[0].Monitoring.State,
+          public_dns_name: args.Instances[0].PublicDnsName,
+          private_ip: args.Instances[0].PrivateIpAddress,
+          private_dns_name: args.Instances[0].PrivateDnsName
+        }
+      }, callback);
+    });
+};
+
+exports.create = function (args, callback) {
+  if (args['ssh-pub-key']) {
+    args.keyPath = args['ssh-pub-key'];
+  }
+
+  exports.getKeys(args)
+    .then(exports.createKey)
+    .then(exports.createInstance)
+    .then(function (args) {
+      args.InstanceId = args.CreatedInstances[0].InstanceId;
+      args.state = 'running';
+      return Promise.resolve(args);
+    })
+    .then(exports.waitForInstanceState)
+    .then(exports.getFilteredInstances)
+    .catch(exports.catch)
+    .then(function (args) {
+      var instance = {
+        name: args.name,
+        ip: args.Instances[0].PublicIpAddress,
+        ssh_key: utils.normalizeKeyPath(args['ssh-key']),
+        ssh_port: '22',
+        user: args.user || 'root',
+        aws: {
+          id: args.InstanceId,
+          size: args.Instances[0].InstanceType,
+          image: args.Instances[0].ImageId,
+          monitoring: args.Instances[0].Monitoring.State,
+          region: args.region || DEFAULT_REGION,
+          public_dns_name: args.Instances[0].PublicDnsName,
+          private_ip: args.Instances[0].PrivateIpAddress,
+          private_dns_name: args.Instances[0].PrivateDnsName
+        }
+      };
+
+      if (_.isFunction(callback)) {
+        callback(instance);
+      }
+    });
+};
+
+exports.destroy = function (instance, callback) {
+  var params = {
+    InstanceId: instance.aws.id,
+    region: instance.aws.region
+  };
+
+  exports.destroyInstance(params)
+    .catch(exports.catch)
+    .then(function (args) {
+      if (_.isFunction(callback)) {
+        callback();
+      }
+    });
+};
+
+exports.shutdown = function (instance, callback) {
+  var params = {
+    InstanceId: instance.aws.id,
+    region: instance.aws.region,
+    state: 'stopped'
+  };
+
+  exports.stopInstance(params)
+    .then(exports.waitForInstanceState)
+    .catch(exports.catch)
+    .then(function (args) {
+      if (_.isFunction(callback)) {
+        callback();
+      }
+    });
+};
+
+exports.reboot = function (instance, callback) {
+  var params = {
+    InstanceId: instance.aws.id,
+    region: instance.aws.region,
+    state: 'running'
+  };
+
+  exports.rebootInstance(params)
+    .then(exports.waitForInstanceState)
+    .then(exports.getFilteredInstances)
+    .catch(exports.catch)
+    .then(function (args) {
+      utils.updateInstance(instance.name, {
+        ip: args.Instances[0].PublicIpAddress,
+        aws: {
+          id: args.InstanceId,
+          size: args.Instances[0].InstanceType,
+          image: args.Instances[0].ImageId,
+          monitoring: args.Instances[0].Monitoring.State,
+          region: instance.aws.region,
+          public_dns_name: args.Instances[0].PublicDnsName,
+          private_ip: args.Instances[0].PrivateIpAddress,
+          private_dns_name: args.Instances[0].PrivateDnsName
+        }
+      }, callback);
+    });
+};
+
+exports.getRegions = function (callback) {
+  exports.describeRegions()
+    .catch(exports.catch)
+    .then(function (args) {
+      _.each(args.Regions, function (region) {
+        region._name = region.RegionName;
+      });
+
+      if (_.isFunction(callback)) {
+        callback(args.Regions);
+      }
+    });
+};
+
+exports.getInstances = function (args, callback) {
+  exports.getFilteredInstances(args)
+    .catch(exports.catch)
+    .then(function (args) {
+      _.each(args.Instances, function (instance) {
+        instance._name = instance.InstanceId;
+      });
+
+      if (_.isFunction(callback)) {
+        callback(args.Instances);
+      }
+    });
+};
+
+// Internal functions
 
 function debugLog(data) {
   if (DEBUG) {
@@ -14,7 +183,7 @@ function debugLog(data) {
   }
 }
 
-var ec2 = function () {
+var ec2 = function (args) {
   if (exports.ec2) {
     return exports.ec2;
   }
@@ -28,22 +197,22 @@ var ec2 = function () {
 
   AWS.config.update({
     accessKeyId: vars.AWS_KEY,
-    secretAccessKey: vars.AWS_SECRET
-  });
-
-  AWS.config.update({
-    region: 'us-east-1'
+    secretAccessKey: vars.AWS_SECRET,
+    region: args.region || DEFAULT_REGION
   });
 
   exports.ec2 = new AWS.EC2();
   return exports.ec2;
 };
 
+// TODO: Get setRegion working. Seems like the endpoint needs to be updated as well.
 exports.setRegion = function (args) {
   args = args || {};
 
   return new Promise(function (resolve) {
-    AWS.config.update({ region: args.region || 'us-east-1' });
+    AWS.config.update({
+      region: args.region || DEFAULT_REGION
+    });
     resolve(args);
   });
 };
@@ -67,7 +236,7 @@ exports.createKey = function (args) {
       PublicKeyMaterial: new Buffer(keyData).toString('base64')
     };
 
-    ec2().importKeyPair(params, function (err, data) {
+    ec2(args).importKeyPair(params, function (err, data) {
       if (err) {
         reject(err);
       } else {
@@ -83,7 +252,7 @@ exports.getKeys = function (args) {
   args = args || {};
 
   return new Promise(function (resolve, reject) {
-    ec2().describeKeyPairs({}, function (err, data) {
+    ec2(args).describeKeyPairs({}, function (err, data) {
       if (err) {
         reject(err);
       } else {
@@ -95,11 +264,11 @@ exports.getKeys = function (args) {
   });
 };
 
-exports.getRegions = function (args) {
+exports.describeRegions = function (args) {
   args = args || {};
 
   return new Promise(function (resolve, reject) {
-    ec2().describeRegions({}, function (err, data) {
+    ec2(args).describeRegions({}, function (err, data) {
       if (err) {
         reject(err);
       } else {
@@ -115,7 +284,7 @@ exports.getImages = function (args) {
   args = args || {};
 
   return new Promise(function (resolve, reject) {
-    ec2().describeImages({
+    ec2(args).describeImages({
       ImageIds: args.ImageIds || [],
       Owners: args.Owners || [],
       Filters: args.Filters || []
@@ -131,19 +300,23 @@ exports.getImages = function (args) {
   });
 };
 
-exports.getInstances = function (args) {
+exports.getFilteredInstances = function (args) {
   args = args || {};
 
   return new Promise(function (resolve, reject) {
-    ec2().describeInstances({
+    ec2(args).describeInstances({
       Filters: args.Filters || [],
-      InstanceIds: args.InstanceIds || [args.InstanceId]
+      InstanceIds: args.InstanceIds || args.InstanceId ? [args.InstanceId] : []
     }, function (err, data) {
       if (err) {
         reject(err);
       } else {
         debugLog(data);
-        args.Instances = data.Reservations[0].Instances;
+        if (data.Reservations[0]) {
+          args.Instances = data.Reservations[0].Instances;
+        } else {
+          args.Instances = [];
+        }
         resolve(args);
       }
     });
@@ -154,8 +327,6 @@ exports.waitForInstanceState = function (args) {
   args = args || {};
   var state = {};
 
-  utils.grey('Waiting for instance state "' + args.state + '"...');
-
   function whileFn() {
     return !(state.Code === args.state || state.Name === args.state);
   }
@@ -163,7 +334,7 @@ exports.waitForInstanceState = function (args) {
   function checkState() {
     return Promise.delay(exports.pollDelay)
       .then(function () {
-        return exports.getInstances(args);
+        return exports.getFilteredInstances(args);
       }).then(function (args) {
         state = args.Instances[0].State;
         return Promise.resolve();
@@ -171,7 +342,6 @@ exports.waitForInstanceState = function (args) {
   }
 
   return utils.promiseWhile(whileFn, checkState).then(function () {
-    utils.success('Done.');
     return Promise.resolve(args);
   });
 };
@@ -181,26 +351,24 @@ exports.createInstance = function (args) {
   var keyPath = utils.normalizeKeyPath(args.keyPath || utils.CONFIG_DIR + '/keys/overcast.key.pub');
   var keyData = fs.readFileSync(keyPath, 'utf8');
 
-  // http://cloud-images.ubuntu.com/releases/14.04/release/
   var params = {
-    ImageId: args.ami || 'ami-018c9568', // Ubuntu 14.04 LTS, 64 bit, EBS
-    InstanceType: args.size || 't1.micro',
+    ImageId: args.image || args.ami || DEFAULT_AMI,
+    InstanceType: args.size || DEFAULT_SIZE,
     KeyName: utils.createHashedKeyName(keyData),
     MinCount: 1,
     MaxCount: 1,
     Monitoring: {
-      Enabled: !!args.monitoring
+      Enabled: utils.argIsTruthy(args.monitoring)
     }
   };
 
   return new Promise(function (resolve, reject) {
-    ec2().runInstances(params, function (err, data) {
+    ec2(args).runInstances(params, function (err, data) {
       if (err) {
         reject(err);
       } else {
         debugLog(data);
         args.CreatedInstances = data.Instances;
-        utils.success('Instance ' + args.CreatedInstances[0].InstanceId + ' created.');
         resolve(args);
       }
     });
@@ -214,7 +382,7 @@ exports.destroyInstance = function (args) {
   };
 
   return new Promise(function (resolve, reject) {
-    ec2().terminateInstances(params, function (err, data) {
+    ec2(args).terminateInstances(params, function (err, data) {
       if (err) {
         reject(err);
       } else {
@@ -232,7 +400,7 @@ exports.rebootInstance = function (args) {
   };
 
   return new Promise(function (resolve, reject) {
-    ec2().rebootInstances(params, function (err, data) {
+    ec2(args).rebootInstances(params, function (err, data) {
       if (err) {
         reject(err);
       } else {
@@ -250,7 +418,7 @@ exports.stopInstance = function (args) {
   };
 
   return new Promise(function (resolve, reject) {
-    ec2().stopInstances(params, function (err, data) {
+    ec2(args).stopInstances(params, function (err, data) {
       if (err) {
         reject(err);
       } else {
@@ -268,7 +436,7 @@ exports.startInstance = function (args) {
   };
 
   return new Promise(function (resolve, reject) {
-    ec2().startInstances(params, function (err, data) {
+    ec2(args).startInstances(params, function (err, data) {
       if (err) {
         reject(err);
       } else {